Refactor styles into seperate files
